Removes use of editor component from Tracks. Deletes editor as it is no longer used.
