+++ src/frontend/screens/SaveTrack/SaveTrackScreen.tsx
@@ -1,10 +1,9 @@
 import React, {useCallback, useMemo} from 'react';
-import {StyleSheet} from 'react-native';
+import {StyleSheet, View} from 'react-native';
 import {defineMessages, useIntl} from 'react-intl';
 import {useNavigationFromRoot} from '../../hooks/useNavigationWithTypes';
 import {useFocusEffect} from '@react-navigation/native';
 import {SaveTrackButton} from './SaveTrackButton';
-import {Editor} from '../../sharedComponents/Editor';
 import {TrackDescriptionField} from './TrackDescriptionField';
 import {HeaderLeft} from './HeaderLeft';
 import {usePreventAndroidBackButton} from '../../hooks/usePreventAndroidBackButton';
@@ -13,6 +12,14 @@ import {PresetCircleIcon} from '../../sharedComponents/icons/PresetIcon';
 import TrackIcon from '../../images/Track.svg';
 import {getTrackDurationAndDistance} from '../../utils/trackMetrics';
 import {usePresetsQuery} from '../../hooks/server/presets';
+import {ScreenContentWithDock} from '../../sharedComponents/ScreenContentWithDock';
+import {LIGHT_GREY} from '../../lib/styles';
+import {PresetView} from '../../sharedComponents/Editor/PresetView';
+import {Divider} from '../../sharedComponents/Divider';
+import {useActiveProject} from '../../contexts/ActiveProjectContext';
+import {useProjectRoleAndDetails} from '../../hooks/useProjectRoleAndDetails';
+import {HeaderText} from '../../sharedComponents/Text/HeaderText';
+import {TrackStats} from '../../sharedComponents/Editor/TrackStats';
 
 export const SaveTrackScreen = () => {
   const navigation = useNavigationFromRoot();
@@ -26,6 +33,8 @@ export const SaveTrackScreen = () => {
     () => Array.from(presets ?? []).filter(p => p.geometry.includes('line')),
     [presets],
   );
+  const {projectId} = useActiveProject();
+  const projectDetails = useProjectRoleAndDetails(projectId);
 
   const canChoosePreset = trackPresets.length > 0;
 
@@ -42,33 +51,55 @@ export const SaveTrackScreen = () => {
   );
 
   return (
-    <Editor
-      presetName={presetName}
-      notesComponent={<TrackDescriptionField />}
-      isTrack={true}
-      trackDurationMs={durationMs}
-      trackDistance={distance}
-      PresetIcon={
-        preset ? (
-          <PresetCircleIcon iconId={preset.iconRef?.docId} size="medium" />
-        ) : (
-          <TrackIcon style={styles.icon} />
-        )
-      }
-      onPressPreset={
-        canChoosePreset
-          ? () =>
-              navigation.navigate('TrackCategoryChooser', {
-                trackAction: 'saveNew',
-              })
-          : undefined
-      }
-      presetDisabled={!canChoosePreset}
-    />
+    <ScreenContentWithDock
+      dockContainerStyle={{padding: 0}}
+      dockContent={undefined}>
+      <View style={styles.container}>
+        <View
+          style={{
+            backgroundColor: projectDetails.projectColor,
+            paddingVertical: 10,
+            borderTopLeftRadius: 12,
+            borderTopRightRadius: 12,
+            paddingHorizontal: 20,
+          }}>
+          <HeaderText variant="header6">
+            {projectDetails.projectHeader}
+          </HeaderText>
+        </View>
+        <PresetView
+          presetName={presetName}
+          PresetIcon={
+            preset ? (
+              <PresetCircleIcon iconId={preset.iconRef?.docId} size="medium" />
+            ) : (
+              <TrackIcon style={styles.icon} />
+            )
+          }
+          onPressPreset={
+            canChoosePreset
+              ? () =>
+                  navigation.navigate('TrackCategoryChooser', {
+                    trackAction: 'saveNew',
+                  })
+              : undefined
+          }
+          presetDisabled={!canChoosePreset}
+        />
+        <Divider />
+        <TrackStats distance={distance} durationMs={durationMs} />
+      </View>
+      <TrackDescriptionField />
+    </ScreenContentWithDock>
   );
 };
 
 const styles = StyleSheet.create({
+  container: {
+    borderRadius: 12,
+    borderWidth: 1,
+    borderColor: LIGHT_GREY,
+  },
   icon: {
     width: 30,
     height: 30,
